+++ otter/assign/utils.py
@@ -196,6 +196,7 @@ def write_otter_config_file(master, result, assignment):
     with open(result / 'student' / config_name, "w+") as f:
         json.dump(config, f, indent=4)
 
+# TODO: update for new assign format
 def run_generate_autograder(result, assignment, args):
     """
     Runs Otter Generate on the autograder directory to generate a Gradescope zip file. Relies on 
@@ -214,44 +215,51 @@ def run_generate_autograder(result, assignment, args):
     os.chdir(str(result / 'autograder'))
     generate_cmd = ["generate", "autograder"]
 
-    if generate_args.get('points', None) is not None:
-        generate_cmd += ["--points", str(generate_args.get('points', None))]
+    # if generate_args.get('points', None) is not None:
+    #     generate_cmd += ["--points", str(generate_args.get('points', None))]
     
-    if generate_args.get('threshold', None) is not None:
-        generate_cmd += ["--threshold", str(generate_args.get('threshold', None))]
+    # if generate_args.get('threshold', None) is not None:
+    #     generate_cmd += ["--threshold", str(generate_args.get('threshold', None))]
     
-    if generate_args.get('show_stdout', False):
-        generate_cmd += ["--show-stdout"]
+    # if generate_args.get('show_stdout', False):
+    #     generate_cmd += ["--show-stdout"]
     
-    if generate_args.get('show_hidden', False):
-        generate_cmd += ["--show-hidden"]
+    # if generate_args.get('show_hidden', False):
+    #     generate_cmd += ["--show-hidden"]
     
-    if generate_args.get('grade_from_log', False):
-        generate_cmd += ["--grade-from-log"]
+    # if generate_args.get('grade_from_log', False):
+    #     generate_cmd += ["--grade-from-log"]
     
-    if generate_args.get('seed', None) is not None:
-        generate_cmd += ["--seed", str(generate_args.get('seed', None))]
+    # if generate_args.get('seed', None) is not None:
+    #     generate_cmd += ["--seed", str(generate_args.get('seed', None))]
 
-    if generate_args.get('public_multiplier', None) is not None:
-        generate_cmd += ["--public-multiplier", str(generate_args.get('public_multiplier', None))]
+    # if generate_args.get('public_multiplier', None) is not None:
+    #     generate_cmd += ["--public-multiplier", str(generate_args.get('public_multiplier', None))]
 
     if generate_args.get('pdfs', {}):
-        pdf_args = generate_args.get('pdfs', {})
+        pdf_args = generate_args.pop('pdfs', {})
         token = APIClient.get_token()
-        generate_cmd += ["--token", token]
-        generate_cmd += ["--course-id", str(pdf_args["course_id"])]
-        generate_cmd += ["--assignment-id", str(pdf_args["assignment_id"])]
+        generate_args['token'] = token
+        generate_args['course_id'] = str(pdf_args['course_id'])
+        generate_args['assignment_id'] = str(pdf_args['assignment_id'])
+        # generate_cmd += ["--token", token]
+        # generate_cmd += ["--course-id", str(pdf_args["course_id"])]
+        # generate_cmd += ["--assignment-id", str(pdf_args["assignment_id"])]
 
         if not pdf_args.get("filtering", True):
-            generate_cmd += ["--unfiltered-pdfs"]
+            generate_args['filtering'] = False
+
+        if not pdf_args.get('pagebreaks', True):
+            generate_args['pagebreaks'] = False
     
     if assignment.is_r:
         generate_cmd += ["-l", "r"]
+        generate_args['lang'] = 'r'
 
     # requirements = assignment.requirements or args.requirements
     # requirements = get_relpath(result / 'autograder', pathlib.Path(requirements))
     if assignment.requirements:
-        requirements = get_relpath(result / 'autograder', pathlib.Path(assignment.requirements))
+        requirements = 'requirements.txt' # get_relpath(result / 'autograder', pathlib.Path(assignment.requirements))
         generate_cmd += ["-r", str(requirements)]
         if assignment.overwrite_requirements or args.overwrite_requirements:
             generate_cmd += ["--overwrite-requirements"]
@@ -260,7 +268,12 @@ def run_generate_autograder(result, assignment, args):
         generate_cmd += assignment.files or args.files
 
     if assignment.variables:
-        generate_cmd += ["--serialized-variables", str(assignment.variables)]
+        # generate_cmd += ["--serialized-variables", str(assignment.variables)]
+        generate_args['serialized_variables'] = str(assignment.variables)
+
+    if generate_args:
+        with open("otter_config.json", "w+") as f:
+            json.dump(generate_args, f, indent=2)
     
     # TODO: change this to import and direct call
     parser = get_parser()