+++ otter/plugins/builtin/__init__.py
@@ -1,0 +1,5 @@
+"""
+Builtin Otter plugins
+"""
+
+from .grade_override import GoogleSheetsGradeOverride