+++ test/test_check.py
@@ -54,7 +54,7 @@ class TestCheck(TestCase):
                         )
 
         # run checker command
-        check_command = check_command = ["check",
+        check_command = ["check",
             TEST_FILES_PATH + "file0.py", 
             "-t", TEST_FILES_PATH + "tests"
         ]
@@ -72,28 +72,25 @@ class TestCheck(TestCase):
                     output.getvalue().strip(), 
                     dedent("""\
                         [0.         0.02002002 0.04004004 0.06006006 0.08008008]
-                        4 of 5 tests passed
+                        q1 passed!
+                        q2 results:
 
-                        Tests passed:
-                            q1 q3 q4 q5 
-
-
-                        Tests failed: 
-                           test/test-check/tests/q2.py
-
-                        Test result:
                         Trying:
                             1 == 1
                         Expecting:
                             False
                         **********************************************************************
-                        Line 2, in test/test-check/tests/q2.py 1
+                        Line 2, in q2 0
                         Failed example:
                             1 == 1
                         Expected:
                             False
                         Got:
-                            True"""), 
+                            True
+
+                        q3 passed!
+                        q4 passed!
+                        q5 passed!"""), 
                     "Did not pass correct tests"
                 )
 
@@ -127,8 +124,8 @@ class TestCheck(TestCase):
                         )
 
         # run checker command
-        check_command = check_command = ["check",
-            TEST_FILES_PATH + "file0.py", 
+        check_command = ["check",
+            TEST_FILES_PATH + "test-nb.ipynb", 
             "-t", TEST_FILES_PATH + "tests"
         ]
         args = parser.parse_args(check_command)
@@ -145,28 +142,25 @@ class TestCheck(TestCase):
                     output.getvalue().strip(), 
                     dedent("""\
                         [0.         0.02002002 0.04004004 0.06006006 0.08008008]
-                        4 of 5 tests passed
+                        q1 passed!
+                        q2 results:
 
-                        Tests passed:
-                            q1 q3 q4 q5 
-
-
-                        Tests failed: 
-                           test/test-check/tests/q2.py
-
-                        Test result:
                         Trying:
                             1 == 1
                         Expecting:
                             False
                         **********************************************************************
-                        Line 2, in test/test-check/tests/q2.py 1
+                        Line 2, in q2 0
                         Failed example:
                             1 == 1
                         Expected:
                             False
                         Got:
-                            True"""), 
+                            True
+
+                        q3 passed!
+                        q4 passed!
+                        q5 passed!"""), 
                     "Did not pass correct tests"
                 )