+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**v2.0.0:**
+
+* Changed granularity of results to be test case-by-test case rather than by file
+* Added ability to list requirements directly in assignment metadata w/out requirements.txt file
+
 **v1.1.4:**
 
 * Fixed `KeyError` when kernelspec unparsable from notebook in Otter Assign